requirements.txt
asgiref==3.2.7
Babel==2.8.0
Django==3.0.6
djangorestframework==3.11.0
django-phonenumber-field==3.0.1
django-widget-tweaks==1.4.5
gunicorn==20.0.4
Markdown==3.2.1
phonenumbers==8.10.21
Pillow==7.1.2
psycopg2-binary==2.8.5
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
